Refactored HomePage.jsx to make Header and Navigation reusable components, added allowed origins to cors in index.js
