+++ client/src/styles/Messages.css
@@ -54,7 +54,6 @@
 
 #messagesContainer {
     margin: 0;
-    padding-bottom: 3rem;
     font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
 }
