+++ server/index.js
@@ -18,6 +18,12 @@ import auth from './utils/auth.js';
 
 dotenv.config();
 
+const allowedOrigins = [
+    'http://localhost:3000',
+    'http://localhost:5173',
+    'https://chat-test-bquw.onrender.com',
+];
+
 async function main() {
     try {
         await connectToDatabase();
@@ -42,7 +48,13 @@ async function startServer() {
         const server = createServer(app);
         const io = new Server(server, {
             cors: {
-                origin: 'http://localhost:5173',
+                origin: (origin, callback) => {
+                    if (allowedOrigins.includes(origin)) {
+                        callback(null, true);
+                    } else {
+                        callback(new Error('Origin not allowed'));
+                    }
+                },
                 methods: ['GET', 'POST']
             },
             connectionStateRecovery: {},
@@ -51,11 +63,17 @@ async function startServer() {
 
         const __dirname = dirname(fileURLToPath(import.meta.url));
 
-        app.use(express.urlencoded({extended: true }));
+        app.use(express.urlencoded({ extended: true }));
         app.use(express.json());
         
         app.use(cors({
-            origin: 'http://localhost:5173',
+            origin: (origin, callback) => {
+                if (allowedOrigins.includes(origin) || !origin) {
+                    callback(null, true);
+                } else {
+                    callback(new Error('Not allowed by CORS'));
+                }
+            },
             methods: ['GET', 'POST'],
             credentials: true
         }));
